refactor sitemap agar jadi route handler saja, untuk skip build
